refactor: move context menus to ui
