--- apps/web/components/context-menu/base.tsx
@@ -1,27 +1,0 @@
-import { ContextMenuProps } from '@/store/types'
-import { BaseContextMenu, Entries } from '@/ui/menu'
-import { MenuItem } from 'types'
-
-export function ContextMenu(props: {
-  items: MenuItem[]
-  contextMenuProps: ContextMenuProps
-  show: boolean
-}) {
-  const { top, left, flowNode } = props.contextMenuProps
-  return props.show ? (
-    <div
-      id="modtree-context-menu"
-      style={{ top, left }}
-      className="absolute z-20"
-    >
-      <BaseContextMenu>
-        <Entries
-          items={props.items}
-          className="py-1 px-1"
-          roundedSelection
-          flowNode={flowNode}
-        />
-      </BaseContextMenu>
-    </div>
-  ) : null
-}
