+++ apps/web/pages/index.tsx
@@ -4,7 +4,7 @@ import ModtreeFlow from '@/components/flow/pane'
 import { HomeLoader } from '@/components/Loader'
 import { useUser } from '@/utils/auth0'
 import { useEffect } from 'react'
-import { ContextMenus } from '@/components/context-menu'
+import { ContextMenus } from '@/ui/menu/context-menu'
 import { ModuleInfoModal, DebugModal } from '@/components/modals'
 import { UserProfileModal } from '@/components/user-profile'
 import { RootSearchBox } from '@/ui/search/module'
