+++ apps/web/ui/menu/context-menu.tsx
@@ -5,7 +5,12 @@ import { Dispatch } from 'redux'
 import type { ReactElement } from 'react'
 import type { MouseEvent } from 'react'
 import type { GraphFlowNode } from '@modtree/types'
-import type { ContextMenuType } from 'types'
+import type { ContextMenuProps, ContextMenuType } from 'types'
+import { useAppSelector } from '@/store/redux'
+import { dashed } from '@/utils/array'
+import { items } from '@/components/menu-items'
+import { MenuItem } from 'types'
+import { Entries } from '@/ui/menu'
 
 /**
  * context:
@@ -29,7 +34,7 @@ export function onContextMenu(
   )
 }
 
-export default function ContextMenu(props: {
+export function BaseContextMenu(props: {
   children: ReactElement[] | ReactElement
   static?: boolean
   className?: string
@@ -51,5 +56,49 @@ export default function ContextMenu(props: {
         </Menu.Items>
       </Menu>
     </div>
+  )
+}
+
+export function ContextMenu(props: {
+  items: MenuItem[]
+  contextMenuProps: ContextMenuProps
+  show: boolean
+}) {
+  const { top, left, flowNode } = props.contextMenuProps
+  return props.show ? (
+    <div
+      id="modtree-context-menu"
+      style={{ top, left }}
+      className="absolute z-20"
+    >
+      <BaseContextMenu>
+        <Entries
+          items={props.items}
+          className="py-1 px-1"
+          roundedSelection
+          flowNode={flowNode}
+        />
+      </BaseContextMenu>
+    </div>
+  ) : null
+}
+
+export function ContextMenus() {
+  const { showContextMenu, contextMenuProps } = useAppSelector(
+    (state) => state.modal
+  )
+  const menus = ['flowNodeContextMenu', 'flowPaneContextMenu']
+  const selected = contextMenuProps.menu
+  return (
+    <>
+      {menus.map((menu, index) => (
+        <ContextMenu
+          key={dashed('modtree-context', menu, index)}
+          show={showContextMenu && selected === menu}
+          contextMenuProps={contextMenuProps}
+          items={items[menu]}
+        />
+      ))}
+    </>
   )
 }
